+++ Websites/Unitystation.org/Wordpress theme/unitystation/footer.php
@@ -12,7 +12,7 @@
                                 <img  src="<?php bloginfo('stylesheet_directory'); ?>/img/built-with-resentment.svg" alt="Build with Resentment">
                             </div>
                             <div class="col-lg-4 flex-center">
-                                <a class="navbar-brand" href="#"><img class="ss13-brand-image" src="<?php bloginfo('stylesheet_directory'); ?>/img/ss13_64.png" alt="Footer Logo"></a>
+                                <a class="navbar-brand" href="#"><img class="ss13-brand-image" src="https://unitystation.org/wp-content/uploads/ss13_64.png" alt="Footer Logo"></a>
                             </div>
                             <div class="col-lg-4 flex-center">
                                 <img src="<?php bloginfo('stylesheet_directory'); ?>/img/contains-technical-debt.svg" alt="Contains Technical Dept">
